Pushing small stuff...

+++ be/report.go
@@ -8,12 +8,13 @@ import (
 )
 
 type ReportArgs struct {
-	Version  string  `json:"version"` // Must be "2.0"
-	Id       string  `json:"id"`      // public key.
+	Version string `json:"version"` // Must be "2.0"
+	Id      string `json:"id"`      // public key.
+	// Team     TeamColor This must be added on db layer.`
 	Latitude float64 `json:"latitude"`
 	Longitue float64 `json:"longitude"`
-	X        int32   `json:"x"`
-	Y        int32   `json:"y"`
+	X        float64 `json:"x"` // 0.0..1.0
+	Y        float64 `json:"y"` // 0.0..1.0
 	Image    []byte  `json:"image"`
 }
